Adding inspect blob commands
