+++ cmd/notation/inspect.go
@@ -14,27 +14,22 @@
 package main
 
 import (
-	"crypto/sha256"
-	b64 "encoding/base64"
-	"encoding/hex"
 	"errors"
 	"fmt"
-	"os"
-	"strconv"
-	"strings"
-	"time"
-
 	"github.com/notaryproject/notation-core-go/signature"
 	"github.com/notaryproject/notation-go/plugin/proto"
 	"github.com/notaryproject/notation-go/registry"
 	cmderr "github.com/notaryproject/notation/cmd/notation/internal/errors"
 	"github.com/notaryproject/notation/cmd/notation/internal/experimental"
+	"github.com/notaryproject/notation/cmd/notation/internal/sharedutils"
 	"github.com/notaryproject/notation/internal/cmd"
 	"github.com/notaryproject/notation/internal/envelope"
 	"github.com/notaryproject/notation/internal/ioutil"
 	"github.com/notaryproject/notation/internal/tree"
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
 	"github.com/spf13/cobra"
+	"os"
+	"strconv"
 )
 
 type inspectOpts struct {
@@ -44,29 +39,6 @@ type inspectOpts struct {
 	outputFormat      string
 	allowReferrersAPI bool
 	maxSignatures     int
-}
-
-type inspectOutput struct {
-	MediaType  string `json:"mediaType"`
-	Signatures []signatureOutput
-}
-
-type signatureOutput struct {
-	MediaType             string              `json:"mediaType"`
-	Digest                string              `json:"digest"`
-	SignatureAlgorithm    string              `json:"signatureAlgorithm"`
-	SignedAttributes      map[string]string   `json:"signedAttributes"`
-	UserDefinedAttributes map[string]string   `json:"userDefinedAttributes"`
-	UnsignedAttributes    map[string]string   `json:"unsignedAttributes"`
-	Certificates          []certificateOutput `json:"certificates"`
-	SignedArtifact        ocispec.Descriptor  `json:"signedArtifact"`
-}
-
-type certificateOutput struct {
-	SHA256Fingerprint string `json:"SHA256Fingerprint"`
-	IssuedTo          string `json:"issuedTo"`
-	IssuedBy          string `json:"issuedBy"`
-	Expiry            string `json:"expiry"`
 }
 
 func inspectCommand(opts *inspectOpts) *cobra.Command {
@@ -137,7 +109,7 @@ func runInspect(command *cobra.Command, opts *inspectOpts) error {
 	if err != nil {
 		return err
 	}
-	output := inspectOutput{MediaType: manifestDesc.MediaType, Signatures: []signatureOutput{}}
+	output := sharedutils.InspectOutput{MediaType: manifestDesc.MediaType, Signatures: []sharedutils.SignatureOutput{}}
 	skippedSignatures := false
 	err = listSignatures(ctx, sigRepo, manifestDesc, opts.maxSignatures, func(sigManifestDesc ocispec.Descriptor) error {
 		sigBlob, sigDesc, err := sigRepo.FetchSignatureBlob(ctx, sigManifestDesc)
@@ -149,40 +121,40 @@ func runInspect(command *cobra.Command, opts *inspectOpts) error {
 
 		sigEnvelope, err := signature.ParseEnvelope(sigDesc.MediaType, sigBlob)
 		if err != nil {
-			logSkippedSignature(sigManifestDesc, err)
+			sharedutils.LogSkippedSignature(sigManifestDesc, err)
 			skippedSignatures = true
 			return nil
 		}
 
 		envelopeContent, err := sigEnvelope.Content()
 		if err != nil {
-			logSkippedSignature(sigManifestDesc, err)
+			sharedutils.LogSkippedSignature(sigManifestDesc, err)
 			skippedSignatures = true
 			return nil
 		}
 
 		signedArtifactDesc, err := envelope.DescriptorFromSignaturePayload(&envelopeContent.Payload)
 		if err != nil {
-			logSkippedSignature(sigManifestDesc, err)
+			sharedutils.LogSkippedSignature(sigManifestDesc, err)
 			skippedSignatures = true
 			return nil
 		}
 
 		signatureAlgorithm, err := proto.EncodeSigningAlgorithm(envelopeContent.SignerInfo.SignatureAlgorithm)
 		if err != nil {
-			logSkippedSignature(sigManifestDesc, err)
+			sharedutils.LogSkippedSignature(sigManifestDesc, err)
 			skippedSignatures = true
 			return nil
 		}
 
-		sig := signatureOutput{
+		sig := sharedutils.SignatureOutput{
 			MediaType:             sigDesc.MediaType,
 			Digest:                sigManifestDesc.Digest.String(),
 			SignatureAlgorithm:    string(signatureAlgorithm),
-			SignedAttributes:      getSignedAttributes(opts.outputFormat, envelopeContent),
+			SignedAttributes:      sharedutils.GetSignedAttributes(opts.outputFormat, envelopeContent),
 			UserDefinedAttributes: signedArtifactDesc.Annotations,
-			UnsignedAttributes:    getUnsignedAttributes(envelopeContent),
-			Certificates:          getCertificates(opts.outputFormat, envelopeContent),
+			UnsignedAttributes:    sharedutils.GetUnsignedAttributes(envelopeContent),
+			Certificates:          sharedutils.GetCertificates(opts.outputFormat, envelopeContent),
 			SignedArtifact:        *signedArtifactDesc,
 		}
 
@@ -214,71 +186,7 @@ func runInspect(command *cobra.Command, opts *inspectOpts) error {
 	return nil
 }
 
-func logSkippedSignature(sigDesc ocispec.Descriptor, err error) {
-	fmt.Fprintf(os.Stderr, "Warning: Skipping signature %s because of error: %v\n", sigDesc.Digest.String(), err)
-}
-
-func getSignedAttributes(outputFormat string, envContent *signature.EnvelopeContent) map[string]string {
-	signedAttributes := map[string]string{
-		"signingScheme": string(envContent.SignerInfo.SignedAttributes.SigningScheme),
-		"signingTime":   formatTimestamp(outputFormat, envContent.SignerInfo.SignedAttributes.SigningTime),
-	}
-	expiry := envContent.SignerInfo.SignedAttributes.Expiry
-	if !expiry.IsZero() {
-		signedAttributes["expiry"] = formatTimestamp(outputFormat, expiry)
-	}
-
-	for _, attribute := range envContent.SignerInfo.SignedAttributes.ExtendedAttributes {
-		signedAttributes[fmt.Sprint(attribute.Key)] = fmt.Sprint(attribute.Value)
-	}
-
-	return signedAttributes
-}
-
-func getUnsignedAttributes(envContent *signature.EnvelopeContent) map[string]string {
-	unsignedAttributes := map[string]string{}
-
-	if envContent.SignerInfo.UnsignedAttributes.TimestampSignature != nil {
-		unsignedAttributes["timestampSignature"] = b64.StdEncoding.EncodeToString(envContent.SignerInfo.UnsignedAttributes.TimestampSignature)
-	}
-
-	if envContent.SignerInfo.UnsignedAttributes.SigningAgent != "" {
-		unsignedAttributes["signingAgent"] = envContent.SignerInfo.UnsignedAttributes.SigningAgent
-	}
-
-	return unsignedAttributes
-}
-
-func formatTimestamp(outputFormat string, t time.Time) string {
-	switch outputFormat {
-	case cmd.OutputJSON:
-		return t.Format(time.RFC3339)
-	default:
-		return t.Format(time.ANSIC)
-	}
-}
-
-func getCertificates(outputFormat string, envContent *signature.EnvelopeContent) []certificateOutput {
-	certificates := []certificateOutput{}
-
-	for _, cert := range envContent.SignerInfo.CertificateChain {
-		h := sha256.Sum256(cert.Raw)
-		fingerprint := strings.ToLower(hex.EncodeToString(h[:]))
-
-		certificate := certificateOutput{
-			SHA256Fingerprint: fingerprint,
-			IssuedTo:          cert.Subject.String(),
-			IssuedBy:          cert.Issuer.String(),
-			Expiry:            formatTimestamp(outputFormat, cert.NotAfter),
-		}
-
-		certificates = append(certificates, certificate)
-	}
-
-	return certificates
-}
-
-func printOutput(outputFormat string, ref string, output inspectOutput) error {
+func printOutput(outputFormat string, ref string, output sharedutils.InspectOutput) error {
 	if outputFormat == cmd.OutputJSON {
 		return ioutil.PrintObjectAsJSON(output)
 	}
@@ -298,13 +206,13 @@ func printOutput(outputFormat string, ref string, output inspectOutput) error {
 		sigNode.AddPair("signature algorithm", signature.SignatureAlgorithm)
 
 		signedAttributesNode := sigNode.Add("signed attributes")
-		addMapToTree(signedAttributesNode, signature.SignedAttributes)
+		sharedutils.AddMapToTree(signedAttributesNode, signature.SignedAttributes)
 
 		userDefinedAttributesNode := sigNode.Add("user defined attributes")
-		addMapToTree(userDefinedAttributesNode, signature.UserDefinedAttributes)
+		sharedutils.AddMapToTree(userDefinedAttributesNode, signature.UserDefinedAttributes)
 
 		unsignedAttributesNode := sigNode.Add("unsigned attributes")
-		addMapToTree(unsignedAttributesNode, signature.UnsignedAttributes)
+		sharedutils.AddMapToTree(unsignedAttributesNode, signature.UnsignedAttributes)
 
 		certListNode := sigNode.Add("certificates")
 		for _, cert := range signature.Certificates {
@@ -322,14 +230,4 @@ func printOutput(outputFormat string, ref string, output inspectOutput) error {
 
 	root.Print()
 	return nil
-}
-
-func addMapToTree(node *tree.Node, m map[string]string) {
-	if len(m) > 0 {
-		for k, v := range m {
-			node.AddPair(k, v)
-		}
-	} else {
-		node.Add("(empty)")
-	}
 }
